requirements.txt
certifi==2021.10.8
dataclasses==0.8
elastic-transport==8.1.2
elasticsearch==6.8.2
PyMySQL==1.0.2
python-dotenv==0.20.0
urllib3==1.26.9
